requirements.txt
aiogram
aiogram-tonconnect
pytoniq
pytoniq-core
redis
pytonapi
sqlalchemy
pydantic-settings
asyncpg
greenlet
alembic
aiocron
black
dedust